% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertReactions}
\alias{convertReactions}
\title{convertReactions}
\usage{
convertReactions(infile, outfile, format = "R")
}
\arguments{
\item{infile}{input file name}

\item{outfile}{output file name}

\item{format}{output code format, should be 'mrgsolve', 'nlmixr2' (or 'rxode'), or 'R'/'deSolve' (default)}
}
\description{
convertReactions
}
\value{
  integer 0 if successful
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
out_file <- tempfile(fileext = ".R")
convertReactions(sbml_file, out_file, format = "R")
#'out_file_mrg <- tempfile(fileext = ".cpp")
#'convertReactions(sbml_file, out_file_mrg, format = "mrgsolve")
#'out_file_rx <- tempfile(fileext = ".R")
#'convertReactions(sbml_file, out_file_rx, format = "nlmixr2")
}
