% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCmtSizes}
\alias{getCmtSizes}
\title{getCmtSizes
Outputs the Sizes of Compartments}
\usage{
getCmtSizes(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getCmtSizes
Outputs the Sizes of Compartments
}
\value{
  \code{NumericVector} of values
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getCmtSizes(model)
}
