% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEventMath}
\alias{getEventMath}
\title{getEventMath
Outputs Model which can be simulated}
\usage{
getEventMath(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getEventMath
Outputs Model which can be simulated
}
\value{
  integer 0 if successful
}
\examples{
sbml_file <- system.file("examples", "sbmlevent.xml", package = "r2sbml")
model <- getModel(sbml_file)
getEventMath(model)
}
