% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\title{getFunctionDefinition
Outputs Model which can be simulated}
\usage{
getFunctionDefinition(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getFunctionDefinition
Outputs Model which can be simulated
}
\value{
  integer 0 if successful
}
\examples{
sbml_file <- system.file("examples", "sbmlfunctiondefinition.xml", package = "r2sbml")
model <- getModel(sbml_file)
getFunctionDefinition(model)
}
