% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNumSpecies}
\alias{getNumSpecies}
\title{getSpeciesNum
Outputs the number of Species in the model}
\usage{
getNumSpecies(input_model)
}
\arguments{
\item{input_model}{input should be an SBML model}
}
\description{
getSpeciesNum
Outputs the number of Species in the model
}
\value{
  integer representing the number of species
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getNumSpecies(model)
}
