% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReactionTable}
\alias{getReactionTable}
\title{getReactionTable
Outputs the Information Table for Reactions}
\usage{
getReactionTable(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getReactionTable
Outputs the Information Table for Reactions
}
\value{
  \code{DataFrame} of information
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getReactionTable(model)
}
