% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertReactions}
\alias{convertReactions}
\title{convertReactions}
\usage{
convertReactions(infile, outfile, format = "R")
}
\arguments{
\item{infile}{input file name}

\item{outfile}{output file name}

\item{format}{output code format, one of 'R'/'deSolve' (default), 'mrgsolve',
'nlmixr2' (or 'rxode'), 'MATLAB', 'Julia' or 'ubiquity'.  An unrecognised format is an error.}
}
\value{
NULL invisibly
}
\description{
convertReactions
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
out_file <- tempfile(fileext = ".R")
convertReactions(sbml_file, out_file, format = "R")
out_file_mrg <- tempfile(fileext = ".cpp")
convertReactions(sbml_file, out_file_mrg, format = "mrgsolve")
out_file_rx <- tempfile(fileext = ".R")
convertReactions(sbml_file, out_file_rx, format = "nlmixr2")
out_file_m <- tempfile(fileext = ".m")
convertReactions(sbml_file, out_file_m, format = "MATLAB")
out_file_jl <- tempfile(fileext = ".jl")
convertReactions(sbml_file, out_file_jl, format = "Julia")
out_file_ub <- tempfile(fileext = ".txt")
convertReactions(sbml_file, out_file_ub, format = "ubiquity")
}
