% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertReactions}
\alias{convertReactions}
\title{convertReactions}
\usage{
convertReactions(infile, outfile, format = "R")
}
\arguments{
\item{infile}{input file name}

\item{outfile}{output file name}

\item{format}{output code format, should be 'MATLAB','mrgsolve',rxode' or 'R' (default)}
}
\description{
convertReactions
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
out_file <- tempfile(fileext = ".R")
convertReactions(sbml_file, out_file, format = "R")
}
