% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertReactions}
\alias{convertReactions}
\title{convertReactions}
\usage{
convertReactions(infile, outfile)
}
\arguments{
\item{infile}{input file name}

\item{outfile}{output file name}
}
\description{
convertReactions
}
