% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{echoSBML}
\alias{echoSBML}
\title{echoSBML}
\usage{
echoSBML(filename)
}
\arguments{
\item{filename}{input file name}
}
\description{
echoSBML
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
echoSBML(sbml_file)
}
