% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCmtNames}
\alias{getCmtNames}
\title{getCmtNames
Outputs the Names of Compartments}
\usage{
getCmtNames(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getCmtNames
Outputs the Names of Compartments
}
\value{
  \code{StringVector} of names
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getCmtNames(model)
}
