% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCmtNames}
\alias{getCmtNames}
\title{getCmtNames
Outputs the Names of Compartments}
\usage{
getCmtNames(input_model)
}
\arguments{
\item{SBMLmodel}{input should be an SBML Model}
}
\description{
getCmtNames
Outputs the Names of Compartments
}
