% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getModel}
\alias{getModel}
\title{getModel
Outputs an external pointer to the Model from SBML file}
\usage{
getModel(filename)
}
\arguments{
\item{filename}{input file name, should end with .xml}
}
\description{
getModel
Outputs an external pointer to the Model from SBML file
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
}
