% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getParameterTable}
\alias{getParameterTable}
\title{getParameterTable
Outputs the Information Table for Paremeters}
\usage{
getParameterTable(input_model)
}
\arguments{
\item{SBMLmodel}{input should be an SBML Model}
}
\description{
getParameterTable
Outputs the Information Table for Paremeters
}
