% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReactionMath}
\alias{getReactionMath}
\title{getReactionMath
Outputs Model which can be simulated}
\usage{
getReactionMath(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getReactionMath
Outputs Model which can be simulated
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getReactionMath(model)
}
