% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReactionTable}
\alias{getReactionTable}
\title{getReactionTable
Outputs the Information Table for Reactions}
\usage{
getReactionTable(input_model)
}
\arguments{
\item{SBMLmodel}{input should be an SBML Model}
}
\description{
getReactionTable
Outputs the Information Table for Reactions
}
