% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRuleMath}
\alias{getRuleMath}
\title{getRuleMath
Outputs Model which can be simulated}
\usage{
getRuleMath(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getRuleMath
Outputs Model which can be simulated
}
