% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSpeciesIC}
\alias{getSpeciesIC}
\title{getspeciesIC
Outputs the Initial Concentrations of Species}
\usage{
getSpeciesIC(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getspeciesIC
Outputs the Initial Concentrations of Species
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getSpeciesIC(model)
}
