% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSpeciesIC}
\alias{getSpeciesIC}
\title{getspeciesIC
Outputs the Initial Concentrations of Species}
\usage{
getSpeciesIC(input_model)
}
\arguments{
\item{SBMLmodel}{input should be an SBML Model}
}
\description{
getspeciesIC
Outputs the Initial Concentrations of Species
}
