% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSpeciesNames}
\alias{getSpeciesNames}
\title{getSpeciesNames
Outputs the Names of Compartments}
\usage{
getSpeciesNames(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getSpeciesNames
Outputs the Names of Compartments
}
\value{
  \code{StringVector} of names
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
model <- getModel(sbml_file)
getSpeciesNames(model)
}
