% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSpeciesNames}
\alias{getSpeciesNames}
\title{getSpeciesNames
Outputs the Names of Compartments}
\usage{
getSpeciesNames(input_model)
}
\arguments{
\item{SBMLmodel}{input should be an SBML Model}
}
\description{
getSpeciesNames
Outputs the Names of Compartments
}
