% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\title{getspeciesTable
Outputs the Information Table for Species}
\usage{
getSpeciesTable(input_model)
}
\arguments{
\item{input_model}{input should be an SBML Model}
}
\description{
getspeciesTable
Outputs the Information Table for Species
}
