% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printSBML}
\alias{printSBML}
\title{printSBML}
\usage{
printSBML(filename)
}
\arguments{
\item{filename}{filename should be the name of the xml file}
}
\description{
printSBML
}
\examples{
sbml_file <- system.file("examples", "sbmlsimple.xml", package = "r2sbml")
printSBML(sbml_file)
}
